+++ src/lib/components/v2/admin/UserEditor.svelte
@@ -2,14 +2,37 @@
 	import { invalidateAll, goto } from '$app/navigation';
 	import { page } from '$app/stores';
 	import { nullifyEmptyStrings, removeNullValues } from '$lib/common/component_utilities';
-	import { FormErrorHandler } from '$lib/common/errors';
+	import { AlertError, displayStandardErrorAlert, FormErrorHandler } from '$lib/common/errors';
 	import { onMount } from 'svelte';
 	import Modal from '$lib/components/common/Modal.svelte';
+	import { sortGroupByNameComparator } from '$lib/common/user_utilities';
+	import SlimSelect from 'slim-select';
 
-	/** @type {import('$lib/types').User} */
+	/** @type {import('$lib/types').User & {group_ids: number[]}} */
 	export let user;
+	/** @type {Array<import('$lib/types').Group>} */
+	export let groups;
 	/** @type {(user: import('$lib/types').User) => Promise<Response>} */
 	export let save;
+
+	/** @type {import('$lib/components/common/StandardErrorAlert.svelte').default|undefined} */
+	let errorAlert = undefined;
+
+	/** @type {number[]} */
+	let groupIdsToAdd = [];
+
+	$: userGroups = user.group_ids
+		.map((id) => groups.filter((g) => g.id === id)[0])
+		.sort(sortGroupByNameComparator);
+
+	$: availableGroups = groups
+		.filter((g) => !user.group_ids.includes(g.id))
+		.filter((g) => !groupIdsToAdd.includes(g.id))
+		.sort(sortGroupByNameComparator);
+
+	$: groupsToAdd = groupIdsToAdd
+		.map((id) => groups.filter((g) => g.id === id)[0])
+		.sort(sortGroupByNameComparator);
 
 	let password = '';
 	let confirmPassword = '';
@@ -59,6 +82,10 @@
 	async function confirmSave() {
 		saving = true;
 		try {
+			const groupsSuccess = await addGroups();
+			if (!groupsSuccess) {
+				return;
+			}
 			if (password) {
 				user.password = password;
 			}
@@ -109,9 +136,114 @@
 		user.slurm_accounts = user.slurm_accounts.filter((_, i) => i !== index);
 	}
 
+	/** @type {Modal} */
+	let addGroupModal;
+	/** @type {SlimSelect|undefined} */
+	let groupsSelector;
+	let addGroupError = '';
+	/** @type {number[]} */
+	let selectedGroupsToAdd = [];
+
+	function openAddGroupModal() {
+		selectedGroupsToAdd = [];
+		addGroupError = '';
+		setSlimSelectOptions(groupsSelector, getGroupSlimSelectOptions());
+		addGroupModal.show();
+	}
+
+	function getGroupSlimSelectOptions() {
+		return availableGroups.map((g) => ({ id: g.id, name: g.name }));
+	}
+
+	async function addGroupToUser() {
+		addGroupError = '';
+		if (selectedGroupsToAdd.length === 0) {
+			addGroupError = 'Group is required';
+			return;
+		}
+
+		groupIdsToAdd = [...groupIdsToAdd, ...selectedGroupsToAdd];
+		selectedGroupsToAdd = [];
+		addGroupModal.hide();
+	}
+
+	/**
+	 * @param {number} groupId
+	 */
+	function removeGroupToAdd(groupId) {
+		groupIdsToAdd = groupIdsToAdd.filter((id) => id !== groupId);
+	}
+
+	async function addGroups() {
+		errorAlert?.hide();
+		if (groupIdsToAdd.length === 0) {
+			return true;
+		}
+
+		const headers = new Headers();
+		headers.set('Content-Type', 'application/json');
+
+		const response = await fetch(`/api/auth/users/${user.id}/`, {
+			method: 'PATCH',
+			credentials: 'include',
+			headers,
+			body: JSON.stringify({
+				new_group_ids: groupIdsToAdd
+			})
+		});
+
+		const result = await response.json();
+
+		if (response.ok) {
+			user = { ...user, group_ids: result.group_ids };
+			groupIdsToAdd = [];
+			return true;
+		}
+
+		errorAlert = displayStandardErrorAlert(new AlertError(result, response.status), 'genericError');
+		return false;
+	}
+
 	onMount(() => {
 		initialSuperuserValue = user.is_superuser;
+		setGroupsSlimSelect();
 	});
+
+	function setGroupsSlimSelect() {
+		const elementId = 'group-select';
+		const selectElement = document.getElementById(elementId);
+		selectElement?.classList.remove('invisible');
+		groupsSelector = new SlimSelect({
+			select: `#${elementId}`,
+			settings: {
+				showSearch: true,
+				allowDeselect: true,
+				ariaLabel: 'Select groups'
+			},
+			events: {
+				afterChange: (selection) => {
+					if (selection.length === 0 || selection[0].value === 'Select...') {
+						selectedGroupsToAdd = [];
+					} else {
+						selectedGroupsToAdd = selection.map((o) => Number(o.value));
+					}
+				}
+			}
+		});
+	}
+
+	/**
+	 * Updates SlimSelect options. This rebuilds the HTML elements and unset the selected value.
+	 * @param {SlimSelect|undefined} select
+	 * @param {Array<{ name: string, id: number|string }>} values
+	 */
+	export function setSlimSelectOptions(select, values) {
+		if (!select) {
+			return;
+		}
+		const options = values.map((p) => ({ text: p.name, value: p.id.toString() }));
+		select.setData([{ text: 'Select...', placeholder: true }, ...options]);
+	}
 </script>
 
 <div id="genericError" />
@@ -315,9 +447,45 @@
 				<span class="invalid-feedback">{$validationErrors['cache_dir']}</span>
 			</div>
 		</div>
+		{#if user.id}
+			<div class="row mb-3 has-validation">
+				<span class="col-sm-3 col-form-label text-end fw-bold">Groups</span>
+				<div class="col-sm-9">
+					<div>
+						{#each userGroups as group}
+							<span class="badge text-bg-light me-2 mb-2 fs-6 fw-normal">{group.name}</span>
+						{/each}
+						{#each groupsToAdd as groupToAdd}
+							<span class="badge text-bg-light me-2 mb-2 fs-6 fw-normal">
+								{groupToAdd.name}
+								<button
+									class="btn btn-link p-0 text-danger text-decoration-none remove-badge"
+									type="button"
+									aria-label="Remove group {groupToAdd.name}"
+									on:click={() => removeGroupToAdd(groupToAdd.id)}
+								>
+									&times;
+								</button>
+							</span>
+						{/each}
+						{#if availableGroups.length > 0}
+							<button class="btn btn-light" type="button" on:click={openAddGroupModal}>
+								<i class="bi bi-plus-circle" />
+								Add group
+							</button>
+						{/if}
+					</div>
+				</div>
+			</div>
+		{/if}
 		<div class="row mb-3">
 			<div class="col-sm-9 offset-sm-3">
-				<input type="submit" value="Save" class="btn btn-primary" disabled={saving} />
+				<button type="submit" class="btn btn-primary" disabled={saving}>
+					{#if saving}
+						<span class="spinner-border spinner-border-sm" role="status" aria-hidden="true" />
+					{/if}
+					Save
+				</button>
 			</div>
 		</div>
 	</form>
@@ -343,4 +511,27 @@
 			<button class="btn btn-primary" on:click={confirmSave}>Confirm</button>
 		</svelte:fragment>
 	</Modal>
+
+	<Modal id="addGroupModal" centered={true} bind:this={addGroupModal} focus={false}>
+		<svelte:fragment slot="header">
+			<h1 class="modal-title fs-5">Add group</h1>
+		</svelte:fragment>
+		<svelte:fragment slot="body">
+			<select id="group-select" class="invisible" class:border-danger={addGroupError} multiple />
+			{#if addGroupError}
+				<span class="text-danger">{addGroupError}</span>
+			{/if}
+			<div id="errorAlert-addGroupModal" class="mt-3" />
+		</svelte:fragment>
+		<svelte:fragment slot="footer">
+			<button class="btn btn-secondary" data-bs-dismiss="modal">Cancel</button>
+			<button class="btn btn-primary" on:click={addGroupToUser}> Add </button>
+		</svelte:fragment>
+	</Modal>
 </div>
+
+<style>
+	.remove-badge {
+		line-height: 0;
+	}
+</style>