+++ src/lib/server/api/auth_api.js
@@ -29,11 +29,13 @@ export async function userAuthentication(fetch, data) {
 /**
  * Fetches user identity
  * @param {typeof fetch} fetch
+ * @param {boolean} groupNames indicates whether to load the list of group names.
  * @returns {Promise<import('$lib/types').User|null>}
  */
-export async function getCurrentUser(fetch) {
+export async function getCurrentUser(fetch, groupNames = false) {
 	logger.debug('Retrieving current user');
-	const response = await fetch(env.FRACTAL_SERVER_HOST + '/auth/current-user/', {
+	const url = `${env.FRACTAL_SERVER_HOST}/auth/current-user/?group_names=${groupNames}`;
+	const response = await fetch(url, {
 		method: 'GET',
 		credentials: 'include'
 	});
@@ -91,17 +93,62 @@ export async function listUsers(fetch) {
  * Fetches a user from the server
  * @param {typeof fetch} fetch
  * @param {number|string} userId
+ * @param {boolean} groupIds indicates whether to load the list of group IDs.
  * @returns {Promise<import('$lib/types').User>}
  */
-export async function getUser(fetch, userId) {
+export async function getUser(fetch, userId, groupIds = true) {
 	logger.debug('Fetching user [user_id=%d]', userId);
-	const response = await fetch(`${env.FRACTAL_SERVER_HOST}/auth/users/${userId}/`, {
+	const response = await fetch(`${env.FRACTAL_SERVER_HOST}/auth/users/${userId}/?group_ids=${groupIds}`, {
 		method: 'GET',
 		credentials: 'include'
 	});
 
 	if (!response.ok) {
 		logger.error('Unable to fetch user [user_id=%d]', userId);
+		await responseError(response);
+	}
+
+	return await response.json();
+}
+
+/**
+ * Fetches the list of groups from the server
+ * @param {typeof fetch} fetch
+ * @param {boolean} userIds indicates whether to load the list of user IDs.
+ * @returns {Promise<Array<import('$lib/types').Group>>}
+ */
+export async function listGroups(fetch, userIds = false) {
+	logger.debug('Fetching groups');
+	const url = `${env.FRACTAL_SERVER_HOST}/auth/group/?user_ids=${userIds}`;
+	const response = await fetch(url, {
+		method: 'GET',
+		credentials: 'include'
+	});
+
+	if (!response.ok) {
+		logger.error('Unable to fetch groups');
+		await responseError(response);
+	}
+
+	return await response.json();
+}
+
+/**
+ * Fetches a group from the server
+ * @param {typeof fetch} fetch
+ * @param {number|string} groupId
+ * @returns {Promise<Array<import('$lib/types').Group>>}
+ */
+export async function getGroup(fetch, groupId) {
+	logger.debug('Fetching group %d', groupId);
+	const url = `${env.FRACTAL_SERVER_HOST}/auth/group/${groupId}/`;
+	const response = await fetch(url, {
+		method: 'GET',
+		credentials: 'include'
+	});
+
+	if (!response.ok) {
+		logger.error('Unable to fetch group %d', groupId);
 		await responseError(response);
 	}